Move user modules and tools into libensemble/tools, adjust some __init__ files to simplify import statements
