+++ libensemble/executors/__init__.py
@@ -1,0 +1,7 @@
+import os
+
+if 'BALSAM_DB_PATH' in os.environ:
+    from libensemble.executors.balsam_executor import BalsamMPIExecutor
+
+from libensemble.executors.executor import Executor
+from libensemble.executors.mpi_executor import MPIExecutor
