+++ libensemble/tests/unit_tests/test_libE_main.py
@@ -60,9 +60,11 @@ def test_manager_exception():
     remove_file_if_exists(hfile_abort)
     remove_file_if_exists(pfile_abort)
 
-    with mock.patch('libensemble.libE.manager_main') as managerMock:
+    with mock.patch('libensemble.libE_manager.manager_main') as managerMock:
         managerMock.side_effect = Exception
-        with mock.patch('libensemble.libE.comms_abort') as abortMock:
+        # Collision between libE.py and libE() (after mods to __init__.py) means
+        #   libensemble.libE.comms_abort tries to refer to the function, not file
+        with mock.patch('libensemble.comms_abort') as abortMock:
             abortMock.side_effect = Exception
             # Need fake MPI to get past the Manager only check and dump history
             with pytest.raises(Exception):
